feat: remove dublicates from a cart page and a cart sidebar components
